+++ src/main/kotlin/gogo/gogoshop/domain/plinko/application/dto/PlinkoDto.kt
@@ -2,6 +2,6 @@ package gogo.gogoshop.domain.plinko.application.dto
 
 data class PlinkoResDto(
     val plinkoId: Long,
-    val ticketPrice: Int,
+    val ticketPrice: Long,
     val ticketQuantity: Int
 )
